Refactor code, add save and restart from checkpoint methods
During this commit, we refactored some parts of the code and added methods to save model, optimizer, and scheduler states into a checkpoint and restart training from the checkpoint. We also enhanced the HiDiscModel class with documentations, removed unnecessary comments from the main function, and corrected import call in resnet_multi_bn model.