Update data loaders and finetuning
This commit includes an update of the data loaders which now incorporate a distributed data sampler. The fine-tuning process was also updated to include additional options for backbones. It includes a change in the num_classes variable in the ft.yaml configuration file. The saving and loading system for checkpoints was also improved to be more flexible and robust.