Updated data loaders and revised finetuning procedure
In this update, the data loaders have been redesigned to incorporate a distributed data sampler. Additionally, the fine-tuning process has been finely tuned to provide better support for different backbones. Another significant change is the alteration in the num_classes variable in the ft.yaml configuration file. Lastly, the system for saving and loading checkpoints has been augmented to increase flexibility and robustness.