Add warmup phase and optimize model compatibility
The update introduces a warmup phase for the attack epsilon rate and optimizes compatibility of models. Now the code supports more types of backbone models in distributed training, including "resnetv2_50_gn", "resnet50_multi_bn", and "resnet50". Moreover, an exception will be raised for unsupported models. The introduction of a warmup phase allows for a more controlled escalation of the attack epsilon rate in the beginning of training.