Add support for single GPU usage and PyTorch Image Models (TIMM)
This commit expands the functionality of the model to support running on a single GPU by adding the `single_gpu` configuration parameter in `init_distributed_mode()`. Also, PyTorch Image Models (TIMM) are now supported with definitions for `resnet50`, `resnetv2_50`, and `wide_resnet50_2` options added to `timm_models.py`. Changes were made to the model's backbone selection process in `eval_knn.py` to accommodate the new models, and a new helper script for plotting results was added.